Targets: provide full list of commands to build
Each target now provides a list of full commands for how they should be
handled. This way it's possible to mix make and non-make commands, and
target specs have full flexibility in how they are specified. This also
simplifies the code in the Build class a little bit, and removes quite
a few of the calls from the Config target class back to the Build class.