Handle SIGINT correctly
Given that subprocess.Popen will not raise an exception on failure, but
rather just provide an return code, the command failure handling needs
to be changed accordingly.

docker run needs to be called with --init so that when we terminate it,
it correctly terminates the process inside the container given they
don't have a controlling terminal.

Closes #48