Extract metadata in a single shot
This reduces the time needed to extract metadata from a docker runtime
by 1 order of magnitude. The metadata extraction script is written in
Perl because it's already included in a Debian base system, which is the
base of all of our Docker images. If we want to be more flexible in
terms of requirements for the images, this will need to be rewritten in
something worse (shell, awk).