metadata: extract ccache/sccache statistics
Currently, this assumes that there are no other concurrent builds using
ccache/sccache. On the other hand, that is the only sane way of using
ccache/sccache and still have meaningful statistics, so there's not much
we can do on our side to mitigate that.

ccache statistics requires ccache 3.7+, because that was the version
that introduced stats output in a machine-parseable format
(--print-stats).

Closes #64