dtbs: handle old kernels with a separate target
This will avoid trying `make dtbs_install` when that's not available. In
that case, `dtbs` will now be skipped, and dtbs will be handled by the
new `dtbs-legacy` target.

With this the support for having "alt_commands" for a target, which was
only actually used by `dtbs`, can be removed. In general, it's cleaner
for targets to just refuse to act when we know they would fail.

Fixes #106