LogParser: ignore log lines where decoding fails
This was found when building a tinyconfig kernel with W=3 in an utf-8
environment.

$ tuxmake -k tinyconfig W=3

Probably the log output got garbled in the final log file by having
several processes writing a lot data in parallel. Ignoring those lines
is not the greatest solution, but at the moment is better than having
TuxMake crash when parsing the log file.