kselftest: run after kselftest-merge, but don't depend on it
This adds support for specifying ordering between targets, without
adding dependencies. If A says "runs_after = B", then *if* both A and B
are requested, than A is guaranteed to be built after B. But if B is not
requested, it's not built at all.

The implementation works like this: when instantiating the targets,
targets listed in "runs_after" targets are instantiated and added to the
list of targets as if they were dependencies, and whether or not they
were added due to ordering only is recorded in
__ordering_only_targets__.  At the end, targets that were added for
ordering only and not explicitly requested are removed.

Signed-off-by: Antonio Terceiro <antonio.terceiro@linaro.org>