docker: reduce definition of `gcc_all` and `clang_all`
They now include _only_ the unversioned toolchain images, and not *all*
gcc-* and clang-* images. This was causing our image builds on CI to
take a long time because the gcc_all and clang_all were building all
those images, which is redundant since we already have jobs for gcc-8,
gcc-9, etc.