metadata: resist against invalid JSON
When extracting metadata from the build environment, it is possible that
the resulting JSON is not valid. This can happen because for example the
process that was writing it got killed before having a chance to finish
writing.

If the input JSON fails to parse, it will be ignored, and saved in the
output metadata under the "invalid_metadata" key.