.gitlab-ci.yml: build ci images inside the gitlab registry
This removes the need to build/publish all the tuxmake images when we
need to update our CI images. We also don't need to publish the images
we use for tuxmake CI together with the other real tuxmake images, so
after this is in place we can stop doing that.