Ensure tuxmake output is deterministic
Basically all we need is ensuring that the tarballs created are
deterministic in terms of file ordering, owner/group and timestamp.
Everything else coming from the kernel build itself should already be
deterministic, unless one is doing, say, randconfig builds, in which
case expecting determinism is silly.

metadata.json is non-deterministic because of the build durations, so we
exclude it from the comparison. Everything else should be bit-by-bit
identical between two builds from the same git commit.