Revert "tuxmake: arch: (mips|ppc|riscv|sparc) change default kernel images"
This reverts commit 929698d8e8af1d5b6b3d6f3cb99b543d297fdc4b. We want to
keep the behavior of building and including the kernel image that is
already built by default in a manual kernel build (`make`).

Being able to boot the default image produced by tuxmake in
tuxrun/tuxtest is nice to have, but is something that needs to be
handled on their side. For example tuxrun needs to know which images it
can boot, and be able to look for them inside the build artifacts.

Signed-off-by: Antonio Terceiro <antonio.terceiro@linaro.org>
Signed-off-by: Anders Roxell <anders.roxell@linaro.org>