docker: Makefile: move image set testing into its own target
Separating the tests into their own target makes it possible to run the
tests for all images in a set locally (`make test-clang_12_all`), and
also will cause the entire build to fail if any image fails their
checks.

The test failure not causing the build to fail is because make commands
are not run with "set -e" by default, i.e. this target succeeds:

	something:
		false; true