rpm: install to interpreter-independent location
The %pyproject_* macros install the files to a python version-specific
location, what requires the RPM to be rebuilt when new python3 versions
come around. This makes it impossible to reuse the same RPM on different
Fedora versions (or different non-Fedora systems that also use RPM and
happen to be on different a python3 version).

Instead, just install the necessary files to /usr/share/tuxmake, and add
a symlink in /usr/bin.

Advantages:

- A single RPM package can be used in any RPM-based system that has
  python3 and perl-JSON-PP.

Disadvantages:

* There will be no Python pre-compilation, so each execution will
  recompile the files. OTOH tuxmake is relatively small, so the overhead
  will be negligible comparing to doing a Linux build.
* TuxMake can't be used as a library without some manual work to add
  /usr/share/tuxmake to sys.path.

Fixes: https://gitlab.com/Linaro/tuxmake/-/issues/150