docker: base image: install tuxmake itself
This allows using tuxmake and its toolchains on e.g. kubernetes
clusters, without having to resort to weird setups as docker-in-docker,
podman-in-docker etc. This way, one can just point to the relevant
tuxmake image, and invoke tuxmake from inside the image using the null
runtime using the toolchain that is already installed in that image.