build: ensure artifacts are copied and metadata is collected
This would already be the case when the build is interrupted while
running an external command, what's a lot more probably since the vast
majority of the build time is in running the actual build commands.

In the off chance that the process gets terminated while running
TuxMake's own Python code, this ensures that the post-build actions such
as copying artifacts that were produced so far, and collecting build
metadata, always happen.

This change moves artifact copying and metadata collection out of the
offline section of the build, but that should not be a problem because
all the code executed is under tuxmake control and not the source tree.