build: drop "to reproduce this build locally:" from log
Depending on the exact options that were provided, one cannot really
just copy and paste the command line to reproduce a build locally. The
build might be using a local container image, a private container
registry, or an environment variable that does not apply to anything
else (like a private network address as a sccache backend).