1.0.0 release
This release marks the introduction of a generic runtime library that other
projects can use to transparently run commands inside containers (or not), just
like TuxMake does. In the future, this features might be extracted into a
separate package, but for now it's part of TuxMake.

New features:

- runtime: metadata: include docker/podman version
- runtime: make runtime reusable outside of TuxMake
- target: config: add support for make:${target} in --kconfig-add
- runtime: run_cmd: add echo parameter to control command line logging
- target: config: add support for interactive make targets

Bug fixes:

- build: environment: don't use the input environment twice