build: fix ccache usage with container runtimes
The transformation in CCACHE_DIR to a bind mounted directory that was
different from the original was previously encapsulated in the
container runtime. But when it was moved to the Build class, the
CCACHE_DIR mapping was done *after* the environment was set, so
CCACHE_DIR would point to the wrong location during the build.

In order to also support the null runtime, we need to just use the
original value of CCACHE_DIR.

Fixes: 96738e76ace1 ("runtime: make runtime reusable outside of TuxMake")