metadata: os: Fallback to /usr/lib/os-release
When building TuxMake in a clean Arch Linux chroot [1], the metadata
operating system tests fail.

This happens because /etc/os-release does not exist in the chroot:

  # ls /etc/os-release
  ls: cannot access '/etc/os-release': No such file or directory

  # ls /usr/lib/os-release
  /usr/lib/os-release

On a Arch Linux system, /etc/os-release is not owned by any package but
/usr/lib/os-release is a part of the filesystem package, which is
require to be installed as part of the base package.

  $ pacman -Qo /etc/os-release
  error: No package owns /etc/os-release

  $ pacman -Qo /usr/lib/os-release
  /usr/lib/os-release is owned by filesystem 2021.12.07-1

This was reported as a bug [2] but it was closed with no apparent follow
up discussion.

Per the os-release specification [3], applications should check for the
existence of /etc/os-release and fallback to /usr/lib/os-release if it
does not exist. While the specification says that /etc/os-release should
be a relative symlink to /usr/lib/os-release, there is no language that
states this is a requirement; it just has to be in one of those
locations.

Modify the metadata commands to test for the existence of
/etc/os-release first, falling back to /usr/lib/os-release. With this
change, all the tests pass in both a clean chroot and an OCI container
on Arch Linux.

[1]: https://wiki.archlinux.org/title/DeveloperWiki:Building_in_a_clean_chroot
[2]: https://bugs.archlinux.org/task/64974
[3]: https://www.freedesktop.org/software/systemd/man/os-release.html

Signed-off-by: Nathan Chancellor <nathan@kernel.org>