.gitlab-ci.yml: Wire up Arch Linux integration tests
shunit2 is not packaged by Arch Linux so we need to download it from the
Arch Linux User Repository (AUR), build it, then install it.

Arch Linux is an x86_64-only distribution, so it does not have
x86_64-linux-gnu binaries available. Create symlinks to the unprefixed
binaries so that the integration tests pass.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>