docker: install libclang-rt-dev on LLVM 14+
Debian recently changed the packaging of the compiler runtime files,
which can result in cryptic build failures when building kernels that
contain an older version of CFI, such as the one present in downstream
Android trees. Dumping the call to clang from the try-run macro reveals:

  clang: error: missing sanitizer ignorelist: '/usr/lib/llvm-15/lib/clang/15.0.7/share/cfi_ignorelist.txt'

These files are now in the libclang-rt-dev package. Install those
packages to gain these files so everything continues to build properly.

Link: https://salsa.debian.org/pkg-llvm-team/llvm-toolchain/-/commit/d24f4d88f13021fca6afe9ce9d087c71362af2cf
Signed-off-by: Nathan Chancellor <nathan@kernel.org>