ci: builds: Skip building gcc all for arc and switch riscv to 12
We do not build gcc-10 for riscv, instead we build gcc-12 for riscv
which should be made the default for gcc all. Similarly gcc-9 for
arc is not being build and there are no latest alternatives to build
for arc. Hence skip building for arc in the case of gcc all.

Signed-off-by: Senthil Kumaran S <senthil.kumaran@linaro.org>